Create new GUI that uses matplotlib standard toolbar, customize toolbar
This commit is a high-level change - it creates a new UI file that will
eventually replace the old one, and outlines the change from using
custom widgets for graph controls to using exclusively the
NavigationToolbar2 functions.  However, these functions are insufficient
or incorrect for the application, so it also begins a subclass of
NavigationToolbar2AggQt that will functions correctly, and defines
overriding methods to where those functions will be implemented.