Develop neural network class, add features to AudioAnalyzer, comment
First, the initial design for the neural network that will be used for
identification is in place.  It is scalable to any number of hidden
layers of any size, can be initialized to accept input or output of any
dimensionality, can handle user-defined nonlinear activation functions.
Backprop is not yet implemented, but forward propagation is.  Everything
uses Numpy to its limit and it seems absurdly efficient.

Changes were also made to the AudioAnalyzer class to account for this
new arrangement, including defining new class variables for storing data
like the assigned classfication and writing headers methods for
converting between the neural network's vector classification and the
analyzer's integer classification.

The comments have also been rewritten to comply with PEP8 and PEP257,
which I had never heard of before.  Some of the comments are yet
unwritten, but all are formatted correctly.