Begin implementation of x-boundedness only and multi-axis plotting
There are only a few changes to these windows, but they mostly revolve
around undoing previous work related to assuming only a single plot and
a single set of axes.  That has been removed - now there can be multiple
axes, all with twinned x-domains but with diverse ranges.  In theory
this will allow me to plot things like an entropy signal on top of a
spectrogram, but we'll see how that goes.  A number of methods have been
removed to account for the changes, or commented out until I can do them
right later.

The comments have also been rewritten to comply with PEP8 and PEP257,
which I had never heard of before.  Some of the comments are yet
unwritten, but all are formatted correctly.