Continue development of the numerical analysis side of the project
The bulk of the work in this commit is around converting between
classification schemes used by the neural network and by the analyzer.
Converting from the analyzer's integer scheme to the neural network's
vectorized scheme is accomplished with a trivial one-hot encoding.
Converting back is done with a windowed moving average where the integer
state selected is the one with the highest probability when all of the
time slices before and after a given point are averaged together using a
specified window.