Implement forward and back button functionality
With this commit I'm starting to switch back to the GUI side, since I
won't be developing my own neural net it's time to get the first GUI
under control.  I need to work on the toolbar buttons and limit checking
for navigation first.  Once that's done I'll proceed with the issue of
applying classifications to data, and after that, with trying to put
multiple plots on different axes in the same figure.