Refactor model and start to implement playback
The biggest change is that all data storage has been offloaded from the
Analyzer class to the SongFile class.  Most processed information (time
array, frequencies array) are held in each instance of the SongFile.
ONLY the full STFT is held in the Analyzer, because that is large enough
that it is not possible to hold one for each song in memory.  AudioGui
has been changed to use the new model.

Additionally, I'm starting to implement playback and the GUI hooks
needed to play part of a SongFile non-blockingly.