Employ signals and slots.  Finally.
Using QtCore's signals and slots mechanism I pulled data that should
never have been locked in the Navigation Bar (but was, because it was
tied to the buttons on the nav bar) OUT of that class.  Now, when the
NavigationBar would've updated a variable (marker, or selection, say),
it instead emits a signal.  This signal can either be received (in this
application, the AudioGUI receives the signals) or it can be ignored, if
you like.