Implement live playback
The GUI changes (handling the play button, the cursor bar, and the time
indicator) are all working, as is the PyAudio stream.  In order to
prevent crashing, the GUI was made so that changes to self.marker are
only displayed if it has been at least 0.05 seconds since the last time
self.marker was displayed.