Change UI design goal to be single panel, single chart
Since the entirety of the model is based on having a single active song
at a time, it doesn't make sense to have two plots each with distinct
data unless we want to try to have two models (pass).  So, all functions
will be on one tab, with a single plot and set of plot controls.
However, 

To that end, I've added actions to the menubar and a table to the left
of the plot, as well as removed the 'current_selection' button (in favor
of the menubar actions).  I've also started changing the implementation
to allow for selecting and loading multiple files at once - might use a
menubar item for this as well, but it currently remains a button. 

The overhaul of the class structure will proceed nonetheless, because
the new structure is just... better.  It makes more sense.