Handle song export; change SongFile init function
SongFile.__init__ now takes only data and Fs as required arguments, with
name, start, and length being optional kwargs.  If start is not
provided, it is zero.  If name is not provided it is empty.  If length
is not provided, it is calculated from Fs and the length of the data.

Song export is done with audiolab.  Songs are exported as PCM16 WAV
files.