Multiple unrelated changes.  See message
First, change the PyDev environment to use virtualenv and reinstalled
all tools as best as possible.

Second, eliminate the use of SndFile/scikit.audiolab in favor of
scipy.io.wavfile, since it really wasn't needed and was an extra pip
install PLUS several extra external dependencies.  Fixed import and
export of WAV files after removing that.

Last, add highpass filtering with a parameter min_freq in the startup
params.  Because real data has sensor noise and it screws massively with
the entropy calculation when you have a powerband at C3 thanks to your
local bass orchestra (it was probably 60Hz, but hard to say).