Add results from another neural net test
Increasing the sample size to 15k per class with a large convolutional
neural net took more time (and I ran it for 5 epochs) but resulted in a
decent result: 0.8075 validation accuracy.

Next I'm going to replace the samples_per_class mechanism - I want it to
train on ALL classified samples and select a subset of the noise samples
to us.