Implement signal-to-noise ratio selection for training; test
The amount of data in one sound file is too large to train on, and
contains mostly non-motif data.  Before, training data was constructed
by assembling a certain number of elements of each class - say, 10000
instances for each class to be learned.  Now, all data labelled with a
non-zero class is deemed 'signal' data, and it is all used in training.
All zero-labeled data is deemed 'noise', and from the large amount of
noise a random subset is selected whose size is a multiplicative factor
of the size of the signal data.

An initial test with this method achieved 0.8710 validation accuracy,
which is promising - plus this method just sounds better.