Added decorators and classes for multithreading; changed init
The new initialization stores a dictionary of signals to slots they
should be connected with, and adds a method for connecting or
disconnecting all signals and slots.  Every slot is now a lambda that
takes an arbitrary number of arguments - this makes it easy to decorate
them and put each existing function inside a framework of steps.