Restructure asynchronous gui calls
Using the decorator eliminates the need to split functions into GUI and
non-GUI dictionaries - I can just decorate the ones that should be
called asynchronously.

All signals will be disconnected when a background thread is running -
stuff will update and you can scroll and what have you, but you can't
access data or call GUI functions because... well, it might explode.  I
don't want to deal with mutexes right now, not that important.

The last thing will be a dictionary or asynchronous function names to
their post-processing clean-up display functions.  Every asynchronous
function has to leave the GUI alone.  But, after you load a bunch of
songs that loading method will want to refresh the table that displays
them.  That can't happen from the loading thread, it has to happen 1)
back in the GUI thread and 2) after the loading thread is DONE done.
For that purpose, the 'end_async_gui_call' method will do all the
cleanup.  It just needs an ordered list of what functions to call once
the thread is done, and it will get that based on the name of the
function running in the thread.  Easy peasy!