Adjust naming scheme for methods in AudioGUI
All methods that are asynchronously decorated end in _async.  Many of
these methods require SOME pre-work done in the GUI thread (a la opening
a pop-up to select a folder or file).  The pre-work for a method that
ends in _async is the callback for that method, the one that is called
from the GUI thread, and ends in _callback instead of _async.