Made method naming structure more uniform
All methods ending in _callback are GUI-interacting functions, so they
must always be called in the GUI thread.  Some of these functions to
nothing but call their paired _async method, but that's the point - all
_async methods have a _callback method and are called only from the
_callback method.  

Shortly, all _callback methods will decorated as @pyqtSlots - apparently
if connecting a signal it's best that the thing it's connected to be
decorated as a slot.

As before, all _async methods are decorated with @async_gui_call