Add documentation to threadsafety and change implementation
Threadstream now uses a QTimer and a cache variable (data) to store all
messages it is given until flush is called and to thread-safely call
flush at least periodically.  The default uses a 100ms period and
therefore emits a signal at least every 100ms and more often if the
flush() method is manually called.